refactor(editor): move file-type icons from vmux_ui into vmux_editor
Relocate vmux_ui::file_icon to vmux_editor::lang_icon; page.rs renders glyphs locally via vmux_ui Icon and a local type_icon backed by lang_icon::lang_logo; drop the file: case from PageIconView and pub mod file_icon from vmux_ui.
