+++ crates/vmux_editor/src/lang_icon.rs
@@ -114,183 +114,3 @@ pub fn lang_logo(ext: &str) -> Option<&'static str> {
         _ => return None,
     })
 }
-
-pub fn ext_of(path: &str) -> String {
-    path.rsplit('/')
-        .next()
-        .unwrap_or("")
-        .rsplit_once('.')
-        .map(|(_, e)| e.to_ascii_lowercase())
-        .unwrap_or_default()
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum FileIcon {
-    Folder,
-    Image,
-    Logo(&'static str),
-    Config,
-    Text,
-    Code,
-    File,
-}
-
-pub fn file_icon_kind(path: &str, is_dir: bool) -> FileIcon {
-    if is_dir {
-        return FileIcon::Folder;
-    }
-    let name = path.rsplit('/').next().unwrap_or("");
-    let ext = ext_of(path);
-    if matches!(
-        ext.as_str(),
-        "png" | "jpg" | "jpeg" | "gif" | "webp" | "svg" | "bmp" | "ico"
-    ) {
-        return FileIcon::Image;
-    }
-    let key = match name {
-        "Dockerfile" => "dockerfile",
-        "CMakeLists.txt" => "cmake",
-        _ => ext.as_str(),
-    };
-    if let Some(d) = lang_logo(key) {
-        return FileIcon::Logo(d);
-    }
-    match ext.as_str() {
-        "ini" | "cfg" | "conf" | "lock" | "env" | "properties" | "editorconfig" => FileIcon::Config,
-        "txt" | "log" | "csv" | "text" => FileIcon::Text,
-        "java" | "vala" | "d" | "ron" => FileIcon::Code,
-        _ if matches!(name, "Makefile" | "makefile" | "GNUmakefile") => FileIcon::Config,
-        _ => FileIcon::File,
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-pub use components::type_icon;
-
-#[cfg(target_arch = "wasm32")]
-mod components {
-    use super::{FileIcon, file_icon_kind};
-    use crate::components::icon::Icon;
-    use dioxus::prelude::*;
-
-    pub fn type_icon(path: &str, is_dir: bool, class: &str) -> Element {
-        match file_icon_kind(path, is_dir) {
-            FileIcon::Folder => folder_glyph(class),
-            FileIcon::Image => image_glyph(class),
-            FileIcon::Logo(d) => logo_icon(d, class),
-            FileIcon::Config => config_glyph(class),
-            FileIcon::Text => text_glyph(class),
-            FileIcon::Code => code_glyph(class),
-            FileIcon::File => file_glyph(class),
-        }
-    }
-
-    fn folder_glyph(class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}",
-                path { d: "M4 20h16a2 2 0 0 0 2-2V8a2 2 0 0 0-2-2h-7.9a2 2 0 0 1-1.69-.9L9.6 3.9A2 2 0 0 0 7.93 3H4a2 2 0 0 0-2 2v13c0 1.1.9 2 2 2Z" }
-            }
-        }
-    }
-
-    fn file_glyph(class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}",
-                path { d: "M15 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7Z" }
-                path { d: "M14 2v4a2 2 0 0 0 2 2h4" }
-            }
-        }
-    }
-
-    fn text_glyph(class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}",
-                path { d: "M15 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7Z" }
-                path { d: "M14 2v4a2 2 0 0 0 2 2h4" }
-                path { d: "M16 13H8" }
-                path { d: "M16 17H8" }
-                path { d: "M10 9H8" }
-            }
-        }
-    }
-
-    fn code_glyph(class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}",
-                path { d: "m16 18 6-6-6-6" }
-                path { d: "m8 6-6 6 6 6" }
-            }
-        }
-    }
-
-    fn config_glyph(class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}",
-                path { d: "M8 3H7a2 2 0 0 0-2 2v5a2 2 0 0 1-2 2 2 2 0 0 1 2 2v5c0 1.1.9 2 2 2h1" }
-                path { d: "M16 21h1a2 2 0 0 0 2-2v-5c0-1.1.9-2 2-2a2 2 0 0 1-2-2V5a2 2 0 0 0-2-2h-1" }
-            }
-        }
-    }
-
-    fn image_glyph(class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}",
-                path { d: "M19 3H5a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2V5a2 2 0 0 0-2-2Z" }
-                path { d: "m21 15-5-5L5 21" }
-            }
-        }
-    }
-
-    fn logo_icon(d: &str, class: &str) -> Element {
-        rsx! {
-            Icon { class: "{class}", fill: "currentColor", stroke: "none",
-                path { d: "{d}" }
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn ext_of_reads_last_segment_extension() {
-        assert_eq!(ext_of("file:///a/b/main.rs"), "rs");
-        assert_eq!(ext_of("/x/Photo.PNG"), "png");
-        assert_eq!(ext_of("/x/noext"), "");
-    }
-
-    #[test]
-    fn rust_file_uses_rust_logo() {
-        assert_eq!(
-            file_icon_kind("file:///x/main.rs", false),
-            FileIcon::Logo(lang_logo("rs").unwrap())
-        );
-    }
-
-    #[test]
-    fn dockerfile_by_name_uses_docker_logo() {
-        assert_eq!(
-            file_icon_kind("/x/Dockerfile", false),
-            FileIcon::Logo(lang_logo("dockerfile").unwrap())
-        );
-    }
-
-    #[test]
-    fn directory_uses_folder() {
-        assert_eq!(file_icon_kind("/x/src", true), FileIcon::Folder);
-    }
-
-    #[test]
-    fn image_uses_image() {
-        assert_eq!(file_icon_kind("/x/a.png", false), FileIcon::Image);
-    }
-
-    #[test]
-    fn text_and_code_and_unknown_fall_back() {
-        assert_eq!(file_icon_kind("/x/notes.txt", false), FileIcon::Text);
-        assert_eq!(file_icon_kind("/x/space.ron", false), FileIcon::Code);
-        assert_eq!(file_icon_kind("/x/data.bin", false), FileIcon::File);
-    }
-}
