+++ crates/vmux_editor/src/lib.rs
@@ -27,6 +27,8 @@ pub use lsp::LspPlugin;
 pub mod page_model;
 
 #[cfg(target_arch = "wasm32")]
+mod lang_icon;
+#[cfg(target_arch = "wasm32")]
 pub mod lsp_page;
 #[cfg(target_arch = "wasm32")]
 pub mod page;
