+++ crates/vmux_ui/src/lib.rs
@@ -9,8 +9,6 @@ pub mod agent_accent;
 
 pub mod favicon;
 
-pub mod file_icon;
-
 pub mod icon;
 
 pub mod theme;
